Moving tests for StrictDataset
